move tests to inline mod
