+++ actix-limitation/tests/mod.rs
@@ -1,19 +1,5 @@
+use actix_limitation::{Error, Limiter};
 use uuid::Uuid;
-
-use super::*;
-
-#[test]
-fn test_create_limiter() {
-    let builder = Limiter::build("redis://127.0.0.1:6379/1");
-    let limiter = builder.finish();
-    assert!(limiter.is_ok());
-
-    let limiter = limiter.unwrap();
-    assert_eq!(limiter.limit, 5000);
-    assert_eq!(limiter.period, Duration::from_secs(3600));
-    assert_eq!(limiter.cookie_name, DEFAULT_COOKIE_NAME);
-    assert_eq!(limiter.session_key, DEFAULT_SESSION_KEY);
-}
 
 #[test]
 #[should_panic = "Redis URL did not parse"]
